# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: eBay Image Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /images/upscale:
    post:
      operationId: upscaleImage
      tags: [images]
      summary: Upscale a single image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpscaleRequest"
      responses:
        "200":
          description: Upscaled image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageResult"
  /images/batch-upscale:
    post:
      operationId: batchUpscaleImages
      tags: [images]
      summary: Batch upscale multiple images
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchUpscaleRequest"
      responses:
        "200":
          description: Batch upscale results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchImageResult"
  /images/convert:
    post:
      operationId: convertImage
      tags: [images]
      summary: Convert image format
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConvertRequest"
      responses:
        "200":
          description: Converted image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageResult"
  /images/metadata:
    post:
      operationId: processMetadata
      tags: [images]
      summary: Read, edit or strip image metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetadataRequest"
      responses:
        "200":
          description: Metadata result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetadataResult"
  /images/banner:
    post:
      operationId: createBanner
      tags: [images]
      summary: Create eBay banner with white background and shipping logo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerRequest"
      responses:
        "200":
          description: Banner image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageResult"
  /images/ai-edit:
    post:
      operationId: aiEditImage
      tags: [images]
      summary: AI-powered text editing on product images
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiEditRequest"
      responses:
        "200":
          description: Edited image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageResult"
...
components:
  schemas:
...
    UpscaleRequest:
      type: object
      properties:
        imageData:
          type: string
          description: Base64 encoded image data (data URL)
        filename:
          type: string
        targetWidth:
          type: integer
          description: Target width in pixels (1600-6500)
        targetHeight:
          type: integer
          description: Target height in pixels (1600-6500)
        outputFormat:
          type: string
          enum: [jpg, jpeg, png, webp, avif, tiff, bmp]
        quality:
          type: integer
          description: Output quality 1-100
        stripMetadata:
          type: boolean
        addEbayOptimization:
          type: boolean
          description: Apply eBay-specific CTR optimization
        useAiEnhancement:
          type: boolean
      required: [imageData, filename]
    BatchUpscaleRequest:
      type: object
      properties:
        images:
          type: array
          items:
            $ref: "#/components/schemas/UpscaleRequest"
        globalSettings:
          $ref: "#/components/schemas/UpscaleRequest"
      required: [images]
    ConvertRequest:
      type: object
      properties:
        imageData:
          type: string
        filename:
          type: string
        outputFormat:
          type: string
          enum: [jpg, jpeg, png, webp, avif, tiff, bmp, gif]
        quality:
          type: integer
        stripMetadata:
          type: boolean
      required: [imageData, filename, outputFormat]
    MetadataRequest:
      type: object
      properties:
        imageData:
          type: string
        filename:
          type: string
        action:
          type: string
          enum: [read, strip, edit]
        metadata:
          type: object
          additionalProperties:
            type: string
          description: Key-value metadata to write when action is edit
      required: [imageData, filename, action]
    MetadataResult:
      type: object
      properties:
        metadata:
          type: object
          additionalProperties: true
        imageData:
          type: string
          description: Processed image (for strip/edit)
        filename:
          type: string
      required: [metadata]
    BannerRequest:
      type: object
      properties:
        imageData:
          type: string
          description: Product image base64
        filename:
          type: string
        country:
          type: string
          description: Shipping country code (US, UK, CA, AU, DE, FR, IT, ES, JP)
          default: US
        logoPosition:
          type: string
          enum: [top-left, top-right, bottom-left, bottom-right]
          default: bottom-right
        bannerWidth:
          type: integer
          default: 1600
        bannerHeight:
          type: integer
          default: 1600
        outputFormat:
          type: string
          enum: [jpg, jpeg, png, webp]
          default: jpg
      required: [imageData, filename]
    AiEditRequest:
      type: object
      properties:
        imageData:
          type: string
        filename:
          type: string
        editInstruction:
          type: string
          description: Natural language instruction (e.g. "change 30 tablets to 60 tablets")
        region:
          type: string
          description: Optional region hint (e.g. "bottom label")
      required: [imageData, filename, editInstruction]
    ImageResult:
      type: object
      properties:
        imageData:
          type: string
          description: Processed image as base64 data URL
        filename:
          type: string
        width:
          type: integer
        height:
          type: integer
        format:
          type: string
        sizeBytes:
          type: integer
        processingTime:
          type: number
      required: [imageData, filename]
    BatchImageResult:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/ImageResult"
        successCount:
          type: integer
        failureCount:
          type: integer
        totalTime:
          type: number
      required: [results, successCount, failureCount, totalTime]
...